+++ integration/testdata/with_custom_function_arbitrary_functions/functions/bulkPersonUpload.ts
@@ -1,0 +1,11 @@
+import { BulkPersonUpload } from "@teamkeel/sdk";
+
+export default BulkPersonUpload(async (value, api, _) => {
+  api.permissions.allow();
+
+  return {
+    people: await Promise.all(
+      value.people.map((p) => api.models.person.create(p))
+    ),
+  };
+});