+++ node/codegen_test.go
@@ -390,7 +390,7 @@ export interface getPerson_input {
 }`
 
 	runWriterTest(t, schema, expected, func(s *proto.Schema, w *Writer) {
-		writeMessages(w, s)
+		writeMessages(w, s, false)
 	})
 }
 
@@ -411,7 +411,7 @@ export interface createPerson_input {
 }`
 
 	runWriterTest(t, schema, expected, func(s *proto.Schema, w *Writer) {
-		writeMessages(w, s)
+		writeMessages(w, s, false)
 	})
 }
 
@@ -432,7 +432,7 @@ export interface createPerson_input {
 }`
 
 	runWriterTest(t, schema, expected, func(s *proto.Schema, w *Writer) {
-		writeMessages(w, s)
+		writeMessages(w, s, false)
 	})
 }
 
@@ -460,7 +460,7 @@ export interface updatePerson_input {
 }`
 
 	runWriterTest(t, schema, expected, func(s *proto.Schema, w *Writer) {
-		writeMessages(w, s)
+		writeMessages(w, s, false)
 	})
 }
 
@@ -496,7 +496,7 @@ export interface listPeople_input {
 }`
 
 	runWriterTest(t, schema, expected, func(s *proto.Schema, w *Writer) {
-		writeMessages(w, s)
+		writeMessages(w, s, false)
 	})
 }
 
@@ -541,7 +541,7 @@ export interface listPeople_input {
 }`
 
 	runWriterTest(t, schema, expected, func(s *proto.Schema, w *Writer) {
-		writeMessages(w, s)
+		writeMessages(w, s, false)
 	})
 }
 
@@ -559,7 +559,7 @@ export interface deletePerson_input {
 }`
 
 	runWriterTest(t, schema, expected, func(s *proto.Schema, w *Writer) {
-		writeMessages(w, s)
+		writeMessages(w, s, false)
 	})
 }
 
@@ -580,7 +580,7 @@ export interface getPersonName_input {
 }`
 
 	runWriterTest(t, schema, expected, func(s *proto.Schema, w *Writer) {
-		writeMessages(w, s)
+		writeMessages(w, s, false)
 	})
 }
 
@@ -606,7 +606,7 @@ export interface GetInput {
 }`
 
 	runWriterTest(t, schema, expected, func(s *proto.Schema, w *Writer) {
-		writeMessages(w, s)
+		writeMessages(w, s, false)
 	})
 }
 
@@ -632,7 +632,7 @@ export interface PersonNameResponse {
 }`
 
 	runWriterTest(t, schema, expected, func(s *proto.Schema, w *Writer) {
-		writeMessages(w, s)
+		writeMessages(w, s, false)
 	})
 }
 
@@ -653,7 +653,7 @@ export interface deletePerson_input {
 }`
 
 	runWriterTest(t, schema, expected, func(s *proto.Schema, w *Writer) {
-		writeMessages(w, s)
+		writeMessages(w, s, false)
 	})
 }
 
@@ -679,7 +679,7 @@ export interface DeleteInput {
 }`
 
 	runWriterTest(t, schema, expected, func(s *proto.Schema, w *Writer) {
-		writeMessages(w, s)
+		writeMessages(w, s, false)
 	})
 }
 
@@ -705,7 +705,7 @@ export interface DeleteResponse {
 }`
 
 	runWriterTest(t, schema, expected, func(s *proto.Schema, w *Writer) {
-		writeMessages(w, s)
+		writeMessages(w, s, false)
 	})
 }
 
@@ -730,7 +730,7 @@ export interface PeopleInput {
 }`
 
 	runWriterTest(t, schema, expected, func(s *proto.Schema, w *Writer) {
-		writeMessages(w, s)
+		writeMessages(w, s, false)
 	})
 }
 
@@ -753,7 +753,7 @@ export interface Foo {
 	`
 
 	runWriterTest(t, schema, expected, func(s *proto.Schema, w *Writer) {
-		writeMessages(w, s)
+		writeMessages(w, s, false)
 	})
 }
 
@@ -792,11 +792,11 @@ export interface Response {
 }`
 
 	runWriterTest(t, schema, inputExpected, func(s *proto.Schema, w *Writer) {
-		writeMessages(w, s)
+		writeMessages(w, s, false)
 	})
 
 	runWriterTest(t, schema, responseExpected, func(s *proto.Schema, w *Writer) {
-		writeMessages(w, s)
+		writeMessages(w, s, false)
 	})
 }
 
@@ -817,7 +817,7 @@ export interface People {
 }`
 
 	runWriterTest(t, schema, expected, func(s *proto.Schema, w *Writer) {
-		writeMessages(w, s)
+		writeMessages(w, s, false)
 	})
 }
 
@@ -846,7 +846,7 @@ export interface Details {
 `
 
 	runWriterTest(t, schema, expected, func(s *proto.Schema, w *Writer) {
-		writeMessages(w, s)
+		writeMessages(w, s, false)
 	})
 }
 
@@ -876,7 +876,7 @@ export interface PersonResponse {
 }`
 
 	runWriterTest(t, schema, expected, func(s *proto.Schema, w *Writer) {
-		writeMessages(w, s)
+		writeMessages(w, s, false)
 	})
 }